# apispecs/swaggerzoom.yml
# swagger.yml

openapi: 3.0.0
info:
  title: "Integration Manager API Spec"
  description: "Integration Manager adapted for Zoom. This swagger interface is using Open API for easier abstraction of APIs for vendors and institutions."
  version: "1.0.0"

servers:
  - url: "/zoomapi"

components:
  securitySchemes:
    AccessKeyAuth:
      type: apiKey
      in: header
      name: X-Access-Key
      x-apikeyInfoFunc: src.zoom_handlers.validate_access_key

security:
  - AccessKeyAuth: []

paths:
  /instructor/recordings:
    get:
      summary: Get recordings for instructor
      operationId: src.zoom_handlers.get_instructor_recordings
      parameters:
        - name: instructor_id
          in: query
          required: true
          schema:
            type: string
          description: Instructor's email or login ID from LTI launch
        - name: course_id
          in: query
          required: true
          schema:
            type: string
          description: Course ID from LTI launch to match against tracking fields
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  recordings:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        uuid:
                          type: string
                        topic:
                          type: string
                        start_time:
                          type: string
                        duration:
                          type: integer
                        recording_files:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              file_type:
                                type: string
                                enum: [MP4, TRANSCRIPT]
                              recording_type:
                                type: string
                              download_url:
                                type: string

  /meeting/{meeting_id}/recordings:
    get:
      summary: Get meeting recordings
      operationId: src.zoom_handlers.get_meeting_recordings
      parameters:
        - name: meeting_id
          in: path
          required: true
          schema:
            type: string
          description: Meeting ID to fetch recordings
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  recordings:
                    type: array
                    items:
                      type: object

  /meeting/{meeting_id}/transcript:
    get:
      summary: Get meeting transcript
      operationId: src.zoom_handlers.get_meeting_transcript
      parameters:
        - name: meeting_id
          in: path
          required: true
          schema:
            type: string
          description: Meeting ID to fetch transcript
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcript:
                    type: array
                    items:
                      type: object
                      properties:
                        index:
                          type: string
                        start:
                          type: string
                        end:
                          type: string
                        text:
                          type: string